Hack to force eventlet not to use kqueue
It seems that on FreeBSD eventlet uses a kqueue (for something). kqueues do not survive a fork, so when something in the child process tries to play with the kqueue, things go terribly wrong and the database ends up quarantined. No good.

Probably the best solution to this problem is to rewrite the updater to not use fork (e.g. use multiprocessing instead, but aint nobody got time for that, so instead we trick eventlet into thinking there is no kqueue support. Seems to work nicely.